feat : 유저 로그인 상태 수정
